fixes
code fixed after ref with installments tests

+++ model/item.go
@@ -4,7 +4,7 @@ type Item struct {
 	SKU         string `json:"sku,omitempty"`
 	Name        string `json:"name,omitempty"`
 	Description string `json:"description,omitempty"`
-	Quantity    int    `json:"quantity,omitempty"`
+	Quantity    int64  `json:"quantity,omitempty"`
 	Price       string `json:"price,omitempty"`
 	Currency    string `json:"currency,omitempty"`
 	Tax         string `json:"tax,omitempty"`
@@ -14,7 +14,7 @@ func NewItem(currency string,
 	description string,
 	name string,
 	price string,
-	quantity int,
+	quantity int64,
 	sku string) *Item {
 	return &Item{
 		Currency:    currency,
